network: quic: Refactor QUIC impl to separate module
